% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cld_contrasts.R
\name{compute_piepho_cld}
\alias{compute_piepho_cld}
\title{Pure Base R Maximal Clique CLD Algorithm (Piepho, 2004)}
\usage{
compute_piepho_cld(
  treatments,
  means,
  p_matrix,
  alpha = 0.05,
  letters = NULL,
  reversed = FALSE
)
}
\description{
Solves the boolean adjacency sweep matrix to generate minimal, standardized grouping letters.
}
\keyword{internal}
