% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiology_gas_exchange.R
\name{get_gam_inflection_points}
\alias{get_gam_inflection_points}
\title{Extract GAM Derivatives and Inflection Points}
\usage{
get_gam_inflection_points(gam_model, time_var = "time", grid_n = 200)
}
\arguments{
\item{gam_model}{A fitted GAM model (\code{mgcv::gam} or similar).}

\item{time_var}{Name of time covariate in model.}

\item{grid_n}{Number of evaluation points across time range (default: 200).}
}
\description{
Evaluates numerical first derivatives (growth rate) and second derivatives (acceleration) 
  from Generalized Additive Models (GAMs) or spline smoothers.
}
