% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_salinity_ions.R
\name{get_membrane_stability_index}
\alias{get_membrane_stability_index}
\title{Calculate Membrane Stability Index and Electrolyte Leakage}
\usage{
get_membrane_stability_index(ec1, ec2)
}
\arguments{
\item{ec1}{Initial electrical conductivity (after 40-50 C or salinity incubation).}

\item{ec2}{Final electrical conductivity (after autoclaving / 100 C boiling).}
}
\value{
A data.frame with Electrolyte Leakage (\eqn{EL \%}) and Membrane Stability Index (\eqn{MSI \%}).
}
\description{
Evaluates cell membrane integrity and thermal/salinity injury from electrical conductivity measurements.
}
