% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endophytes_microbiome.R
\name{get_microbial_synergy}
\alias{get_microbial_synergy}
\title{Wrapper for Microbial Factorial Synergy}
\usage{
get_microbial_synergy(y_control, y_inoc_a, y_inoc_b, y_dual, y_max = NULL)
}
\description{
Wrapper for Microbial Factorial Synergy
}
