% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_salinity_ions.R
\name{get_osmotic_adjustment}
\alias{get_osmotic_adjustment}
\title{Calculate Full Active Osmotic Adjustment (OA100)}
\usage{
get_osmotic_adjustment(
  psi_s_stress,
  rwc_stress,
  psi_s_control,
  rwc_control = 0.95,
  awf = 0.1
)
}
\arguments{
\item{psi_s_stress}{Osmotic potential under stress in MPa (negative number).}

\item{rwc_stress}{Relative Water Content under stress (fraction 0 to 1).}

\item{psi_s_control}{Osmotic potential under well-watered control in MPa (negative number).}

\item{rwc_control}{Relative Water Content under control (default: \code{0.95}).}

\item{awf}{Apoplastic water fraction (default: \code{0.10}).}
}
\value{
A data.frame containing normalized osmotic potentials and active osmotic adjustment (\eqn{\Delta OA_{100}}).
}
\description{
Computes true active osmotic adjustment normalized to 100% relative water content 
  corrected for apoplastic water fraction (\eqn{AWF}).
}
