% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_drc.R
\name{get_relative_potency}
\alias{get_relative_potency}
\title{Extract Relative Potency between Dose-Response Curves}
\usage{
get_relative_potency(model, comp_level = NULL)
}
\arguments{
\item{model}{A fitted multi-curve drc object.}

\item{comp_level}{Benchmark reference curve name.}
}
\description{
Computes relative potency (ratio of ED50 values) between treatments or genotypes.
}
