% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_salinity_ions.R
\name{stress_salinity_ions}
\alias{stress_salinity_ions}
\title{Salinity, Ion Homeostasis, and Osmotic Adjustment Kinetics}
\description{
Potassium/Sodium selectivity coefficients (S_K,Na), full Active Osmotic Adjustment (OA100), 
  and Cell Membrane Stability Index (MSI) / Electrolyte Leakage under salinity stress.
}
