% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_thermal_anthesis.R
\name{stress_thermal_anthesis}
\alias{stress_thermal_anthesis}
\title{Thermal Stress at Anthesis (Z65) and Non-Linear Grain Filling Kinetics}
\description{
Model pollen viability (Binomial GLMM), Cell Membrane Stability (CMS), 
  and Richards/Schnute non-linear grain filling rate (GFR) and duration (GFD).
}
