% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_export.R
\name{theme_agriinsight}
\alias{theme_agriinsight}
\title{Publication ggplot2 Themes for Agri-Biological Sciences}
\usage{
theme_agriinsight(theme = c("academic_light", "academic_dark"))
}
\arguments{
\item{theme}{Theme aesthetic: \code{"academic_light"} or \code{"academic_dark"}.}
}
\description{
Clean, high-contrast ggplot2 theme styled for Nature Plants, Crop Science, and TAG.
}
